TS-29324 Use AgentBase for revision and commitDescriptor endpoint
