+++ agent/src/main/java/com/teamscale/jacoco/agent/testimpact/TestwiseCoverageAgent.java
@@ -10,19 +10,16 @@ import com.squareup.moshi.Moshi;
 import com.squareup.moshi.Types;
 import com.teamscale.client.ClusteredTestDetails;
 import com.teamscale.client.StringUtils;
-import com.teamscale.client.TeamscaleServer;
 import com.teamscale.jacoco.agent.AgentBase;
 import com.teamscale.jacoco.agent.JacocoRuntimeController.DumpException;
 import com.teamscale.jacoco.agent.options.AgentOptions;
 import com.teamscale.report.testwise.jacoco.JaCoCoTestwiseReportGenerator;
 import com.teamscale.report.testwise.jacoco.cache.CoverageGenerationException;
-import com.teamscale.report.testwise.model.RevisionInfo;
 import com.teamscale.report.testwise.model.TestExecution;
 import spark.Request;
 import spark.Response;
 import spark.Service;
 
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -46,10 +43,6 @@ public class TestwiseCoverageAgent extends AgentBase {
 	/** JSON adapter for test executions. */
 	private final JsonAdapter<TestExecution> testExecutionJsonAdapter = new Moshi.Builder().build()
 			.adapter(TestExecution.class);
-
-	/** JSON adapter for revision information. */
-	private final JsonAdapter<RevisionInfo> revisionInfoJsonAdapter = new Moshi.Builder().build()
-			.adapter(RevisionInfo.class);
 
 	/** JSON adapter for test details. */
 	private final JsonAdapter<List<ClusteredTestDetails>> clusteredTestDetailsAdapter = new Moshi.Builder().build()
@@ -78,8 +71,6 @@ public class TestwiseCoverageAgent extends AgentBase {
 	protected void initServerEndpoints(Service spark) {
 		super.initServerEndpoints(spark);
 		spark.get("/test", (request, response) -> controller.getSessionId());
-		spark.get("/revision", (request, response) -> this.getRevisionInfo());
-		spark.put("/revision", this::handleSetRevision);
 		spark.post("/test/start/" + TEST_ID_PARAMETER, this::handleTestStart);
 		spark.post("/test/end/" + TEST_ID_PARAMETER, this::handleTestEnd);
 		spark.post("/testrun/start", this::handleTestRunStart);
@@ -195,24 +186,5 @@ public class TestwiseCoverageAgent extends AgentBase {
 			logger.error("Failed to store test execution: " + e.getMessage(), e);
 			return Optional.empty();
 		}
-	}
-
-	/** Returns revision information for the Teamscale upload. */
-	private String getRevisionInfo() {
-		TeamscaleServer server = options.getTeamscaleServerOptions();
-		return revisionInfoJsonAdapter.toJson(new RevisionInfo(server.commit, server.revision));
-	}
-
-	/** Handles setting the revision. */
-	private String handleSetRevision(Request request, Response response) {
-		String revision = org.conqat.lib.commons.string.StringUtils.removeDoubleQuotes(request.body());
-		if (revision == null || revision.isEmpty()) {
-			return handleBadRequest(response,
-					"The new revision name is missing in the request body! Please add it as plain text.");
-		}
-		logger.debug("Changing revision name to " + revision);
-		options.getTeamscaleServerOptions().revision = revision;
-		response.status(HttpServletResponse.SC_NO_CONTENT);
-		return "";
 	}
 }
